level-1-beginner/requirements.txt
groq
supabase
openai
python-dotenv
tavily-python